Split predictor tests; document model requirements
